+++ tests/core/test_eventlog.rs
@@ -21,7 +21,7 @@ fn test_git_v2_31_events() -> anyhow::Result<()> {
     let repo = git.get_repo()?;
     let conn = repo.get_db_conn()?;
     let event_log_db = EventLogDb::new(&conn)?;
-    let event_replayer = EventReplayer::from_event_log_db(&event_log_db)?;
+    let event_replayer = EventReplayer::from_event_log_db(&repo, &event_log_db)?;
     let events: Vec<Event> = get_event_replayer_events(&event_replayer)
         .iter()
         .cloned()